pyproject.toml
[tool.poetry]
name = "rm_sys_django"
version = "0.1.0"
description = ""
authors = ["rmpeter0474"]

[tool.poetry.dependencies]
python = "3.8.10"
Django = "3.2.5"
djangorestframework = "3.12.4"
msal = "1.10.0"
pandas = "1.2.4"
pycparser = "2.20"
PyJWT = "2.0.1"
python-dateutil = "2.8.1"
pytz = "2021.1"
sqlparse = "0.4.1"
urllib3 = "1.26.4"
pymongo = "3.12.0"
uWSGI = "2.0.20"
django-extensions = "3.1.3"
django-allauth = "0.45.0"
django-debug-toolbar = "3.2.2"
django-redis = "4.7.0"
channels = "3.0.4"
rich = "10.12.0"
asgiref = "3.3.4"
channels-redis = "3.3.1"
djangorestframework-api-key = "2.1.0"
Twisted = {extras = ["http2", "tls"], version = "^22.4.0"}
mysql-connector-python = "8.0.28"
numpy = "^1.22.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
